cleaned up types
Removed types.ts. The required types from that file were movced to ObservationQueryParameters.ts.
